Add basic swap method to linked list
The basic swap method is added to the linked list (this will be useful for performing bubble sort later?).

There is also an additional swap(Transition, Transition) method that I've added. This one is only used for testing purposes (to test if the basic swap method works properly). This can be removed in the future if needed.